break out node specific registry to avoid webpack errors
