+++ bin/jexi.js
@@ -2,10 +2,11 @@
 
 import { jexiInterpreter } from '../src/index.js'
 import { startRepl } from '../src/repl.js'
+import nodeSpecificExtensions from '../src/non-browser.js'
 
 // if they gave a filename:
 if (process.argv.length > 2 && !process.argv[2].startsWith('-')) {
-  const jexi = jexiInterpreter({}, { trace: false })
+  const jexi = jexiInterpreter(nodeSpecificExtensions, { trace: false })
 
   // run it
   const result = await jexi.evaluate({ $run: process.argv[2] })
@@ -15,8 +16,8 @@ if (process.argv.length > 2 && !process.argv[2].startsWith('-')) {
   // eslint-disable-next-line no-console
   console.log('Starting Jexi REPL...')
 
-  // imagining the repl will have it's own custom commands later on:
-  const extensions = {}
+  // imagining the repl may have additional commands (as extensions) later on:
+  const extensions = nodeSpecificExtensions
 
   const jexi = jexiInterpreter(extensions, { trace: false })
